Testing alignment

+++ mean_extraction/alignment_testing.py
@@ -11,14 +11,16 @@
 
 import os
 import ConfigParser
+import random
 import hgbrian_GPU_NW
+import numba
 
 FN_TMP_SEQS = os.path.expanduser('~/facs-seq_test/seqs_tmp.txt')
+FN_TMP_SEQS2 = os.path.expanduser('~/facs-seq_test/seqs_tmp2.txt')
 FN_TMP_SCORES = os.path.expanduser('~/facs-seq_test/scores_tmp.txt')
 MAX_GAPS = 5
-CHUNK_SIZE = 50000
+CHUNK_SIZE = 500000
 THREADS_PER_BLOCK = 64
-
 
 # given a file of \n-delimited seqs and a reference sequence seq_in (a string),
 # get a list where entry i is the NW distance between seq_in and line i in fn_in
@@ -29,11 +31,15 @@ def file_vs_ref_dist(fn_in, seq_in, fn_tmp_seqs = FN_TMP_SEQS, fn_tmp_scores = F
       ft.write(seq_in + '\n')
 
   hgbrian_GPU_NW.score_file(fn_tmp_seqs, fn_tmp_scores, max_gaps, chunk_size, threads_per_block)
+  # clean up a little
+  context = numba.cuda.current_context()
+  context.reset()
+
   ans = []
   with open(fn_tmp_scores, 'r') as fi: # record every other line, starting with the first (line 0)
     for (i,l) in enumerate(fi):
       if i % 2 == 0:
-        ans.append(l)
+        ans.append(float(l.strip()))
   os.remove(fn_tmp_seqs)
   os.remove(fn_tmp_scores)
   return(ans)
@@ -43,7 +49,8 @@ def list_vs_ref_dist(seqs_target_list, seq_in, fn_in_tmp):
     for s in seqs_target_list:
       fo.write(s + '\n')
   ans = file_vs_ref_dist(fn_in_tmp, seq_in)
-  os.remove(fn_in_tmp)
+  if os.path.isfile(fn_in_tmp):
+    os.remove(fn_in_tmp)
   return(ans)
 
 def list_vs_self_dist(seqs_target_list, fn_in_tmp):
@@ -67,25 +74,64 @@ def get_biggest_groups(fn_in, aln_thresh):
           ans.append(gp)
   return(ans)
 
-# given a list of reads, get the longest distance within the reads
-def longest_dist_within_group(reads, fn_in_tmp):
+# given a CSV with header Group,Pre,Ambig,N,Aln, get 'num_out' randomly chosen group IDs
+def get_random_groups(fn_in, num_out, aln_thresh = 0):
+  gps = get_biggest_groups(fn_in, aln_thresh)
+  random.shuffle(gps)
+  return(gps[:num_out])
+
+# for min_score_within_group: get the distances all within one call to score_file
+def list_vs_self_dist_onerun(reads, fn_tmp_seqs = FN_TMP_SEQS, fn_tmp_scores = FN_TMP_SCORES, max_gaps = MAX_GAPS, chunk_size = CHUNK_SIZE, threads_per_block = THREADS_PER_BLOCK):
+
+  with open(fn_tmp_seqs, 'w') as fo:
+    for i in range(len(reads)):
+      for j in range(i+1, len(reads)):
+        fo.write(reads[i] + '\n')
+        fo.write(reads[j] + '\n')
+
+  hgbrian_GPU_NW.score_file(fn_tmp_seqs, fn_tmp_scores, max_gaps, chunk_size, threads_per_block)
+  # clean up a little
+  context = numba.cuda.current_context()
+  context.reset()
+
+  ans = 100000.
+  with open(fn_tmp_scores, 'r') as fi: # record every other line, starting with the first (line 0)
+    for l in fi:
+      l = float(l.strip())
+      if l < ans:
+        ans = l
+  os.remove(fn_tmp_seqs)
+  os.remove(fn_tmp_scores)
+  return(ans)
+
+# given a list of reads, get the minimum score within the reads
+def min_score_within_group(reads, fn_in_tmp):
   '''reads = []
   with open(clustered_read_fn, 'r') as fi:
     for l in fi:
       [seq, _, this_gp_id] = l.strip().split(',')
       if int(this_gp_id) == target_gp_id:
         reads.append(seq)'''
-  dist_dict = list_vs_self_dist(reads, fn_in_tmp)
-  max_dist = -1
-  for k in dist_dict.keys():
-    tmp = max(dist_dict[k])
-    if tmp > max_dist:
-      max_dist = tmp
-  return(max_dist)
+  print len(reads)
+  # remove duplicates
+  reads = list(set(reads))
+  min_score = list_vs_self_dist_onerun(reads, fn_in_tmp)
+  print min_score
+  #dist_dict = list_vs_self_dist(reads, fn_in_tmp)
+  #min_score = 100000
+  #for k in dist_dict.keys():
+  #  d = dist_dict[k]
+  #  if len(d) > 0: # can't take min of an empty sequence
+  #    tmp = min(dist_dict[k])
+  #  else:
+  #    tmp = min_score
+  #  if tmp < min_score:
+  #    min_score = tmp
+  return(min_score)
 
-# given a seq and the filename of a read table, get the minimum distance to another group, and a list of all group IDs with that distance
+# given a seq and the filename of a read table, get the maximum score to another group, and a list of all group IDs with that distance
 # read table has no header; first CSV field is sequence, next is group ID
-def min_dist_in_read_table_to_seq(seq, read_table_fn, fn_in_tmp):
+def max_score_in_read_table_to_seq(seq, read_table_fn, fn_in_tmp):
   seqs = []
   gp_ids = []
   with open(read_table_fn, 'r') as fi:
@@ -96,11 +142,11 @@ def min_dist_in_read_table_to_seq(seq, read_table_fn, fn_in_tmp):
         seqs.append(s)
         gp_ids.append(g)
   
-  # get the distance to each line in the read table
+  # get the score to each line in the read table
   dists = list_vs_ref_dist(seqs, seq, fn_in_tmp)
-  target_dist = min(dists)
+  target_dist = max(dists)
   
-  # get all group IDs with this distance  
+  # get all group IDs with this distance
   gps_out = []
   for (g,d) in zip(gp_ids, dists):
     if d == target_dist:
@@ -115,18 +161,19 @@ def get_seqs_from_group_ids(group_ids, read_table_fn):
     ans_dict[g] = None
   with open(read_table_fn, 'r') as fi:
     for l in fi:
-      [s, g] = l.strip().split(',')[:2]
-      if int(g) in ans.keys():
+      [s, g] = l.strip().split(',')[:2]; g = int(g)
+      if g in ans_dict.keys():
         ans_dict[g] = s
   ans = [ans_dict[q] for q in group_ids]
   return(ans)
-  
+
 # given a config object, get the following:
 # a table gp_id, longest internal distance for all groups above a certain number of reads
 # a 'table' gp_id, # reads, minimum distance to another group, comma-separated list of gp_ids with that distance, for all groups above a certain number of reads
 
-def main_method(cfg, reads_thresh, internal_dist_fn_out, intergroup_dist_fn_out, fn_in_tmp = FN_TMP_SEQS):
-  target_groups = get_biggest_groups(fn_in = os.path.expanduser(cfg.get('Output', 'file_fates')), aln_thresh = reads_thresh)
+def main_method(cfg, num_groups, internal_dist_fn_out, intergroup_dist_fn_out, fn_in_tmp = FN_TMP_SEQS2):
+  #target_groups = get_biggest_groups(fn_in = os.path.expanduser(cfg.get('Output', 'file_fates')), aln_thresh = reads_thresh)
+  target_groups = get_random_groups(fn_in = os.path.expanduser(cfg.get('Output', 'file_fates')), num_out = num_groups)
   target_seqs = get_seqs_from_group_ids(group_ids = target_groups, read_table_fn = os.path.expanduser(cfg.get('Output', 'file_aligned_filtered')))
   # drop groups without a seq found (meaning seq was dropped in filtering)
   final_groups = []
@@ -152,19 +199,22 @@ def main_method(cfg, reads_thresh, internal_dist_fn_out, intergroup_dist_fn_out,
   # modify longest_dist_within_group!
   with open(internal_dist_fn_out, 'w') as fo:
     for g in final_groups:
-      d = longest_dist_within_group(reads_dict[g], fn_in_tmp)
+      print('group: ' + str(g))
+      d = min_score_within_group(reads_dict[g], fn_in_tmp)
       fo.write(str(g) + ',' + str(d) + '\n')
-  
+
   # write a table gp_id, shortest distance to another seq, CSV list of group ids with that distance
   with open(intergroup_dist_fn_out, 'w') as fo:
     for (g,s) in zip(final_groups, final_seqs):
-      target_dist, gps_out = min_dist_in_read_table_to_seq(s, os.path.expanduser(cfg.get('Output', 'file_aligned_filtered')), fn_in_tmp)
-    ans = str(g) + ',' + str(target_dist) + ',' + ','.join(gps_out)
-    fo.write(ans + '\n')
+      target_dist, gps_out = max_score_in_read_table_to_seq(s, os.path.expanduser(cfg.get('Output', 'file_aligned_filtered')), fn_in_tmp)
+      print g
+      print target_dist
+      ans = str(g) + ',' + str(target_dist) + ',' + ','.join([str(q) for q in gps_out])
+      fo.write(ans + '\n')
 
 if __name__ == '__main__': # debug: change this!
   cfg = ConfigParser.RawConfigParser(allow_no_value=True)
   cfg.read(os.path.expanduser('~/facs-seq/GPD/miseq/build_table_align_GPD.cfg'))
-  main_method(cfg, reads_thresh = 50, 
+  main_method(cfg, 10,
     internal_dist_fn_out = os.path.expanduser('~/facs-seq_test/internal_test.csv'), 
     intergroup_dist_fn_out = os.path.expanduser('~/facs-seq_test/intergroup_test.csv'))
